pyproject.toml
[tool.poetry]
name = "ai-bestie"
version = "0.1.0"
description = "AI Bestie application"
authors = ["Moe Abdalla <moe@risolto.ltd>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = ">=0.1.0"
langchain-groq = ">=0.0.1"
langchain-openai = ">=0.1.0"
langchain-anthropic = ">=0.1.0"
pydantic = ">=2.0.0"
python-dotenv = ">=1.0.0"
fastapi = ">=0.109.0"
aiohttp = ">=3.9.0"
pydantic-settings = ">=2.1.0"
uvicorn = ">=0.27.0"
redis = {version = ">=4.3.4", extras = ["hiredis"]}
python-telegram-bot = ">=20.8"
langgraph = ">=0.0.18"
langgraph-checkpoint-sqlite = "^1.0.0"
SQLAlchemy = ">=2.0.0"
tavily-python = "^0.3.3"
aiosqlite = "^0.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
